+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 - [GH-61](https://github.com/zackad/prettier-plugin-twig/issues/61#issuecomment-2596726423) Add support for twig comment as html element attribute
+- [GH-116](https://github.com/zackad/prettier-plugin-twig/issues/116) Add possibility to register any custom Twig test expression
 
 ### Bugfixes
 - Fix broken embed statement with `only` modifier